fix: duplicate database

+++ database/migrations/2025_12_12_113432_create_user_badges_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('user_badges', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->foreignId('badge_id')->constrained()->onDelete('cascade');
-            $table->timestamp('awarded_at');
-            $table->unique(['user_id', 'badge_id']);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('user_badges')) {
+            Schema::create('user_badges', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->foreignId('badge_id')->constrained()->onDelete('cascade');
+                $table->timestamp('awarded_at');
+                $table->unique(['user_id', 'badge_id']);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
